+++ H264Decoder/Data/NALU/PredWeightTable.h
@@ -8,8 +8,8 @@ struct NALUnit;
 
 struct PredWeightTable
 {
-    std::uint8_t lumaLog2WeightDenom = 0;
-    std::uint8_t chromaLog2WeightDenom = 0;
+    std::uint16_t lumaLog2WeightDenom = 0;
+    std::uint16_t chromaLog2WeightDenom = 0;
 
     std::vector<WeightData> weightDataL0;
     std::vector<WeightData> weightDataL1;
